Fixed some includes, moved the project to src/ directory. Added a CMakeLists.txt
Updated README with instructions to build on Unix.
