+++ .gitignore
@@ -2,3 +2,10 @@
 \.vs
 \Debug
 \src\Debug
+.vscode/
+database.db
+cmake_install.cmake
+CMakeCache.txt
+bin/
+CMakeFiles
+build/
