+++ src/core/Models/Segment.cpp
