+++ src/core/Models/Segment.hpp
